+++ backend/app/models/proposal.py
@@ -1,8 +1,9 @@
 # @AI-HINT: Proposal model - freelancer bids on projects with pricing and cover letter
-from sqlalchemy import String, Integer, Float, DateTime, Text, ForeignKey, Boolean
+from sqlalchemy import String, Integer, Float, DateTime, Text, ForeignKey, Boolean, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.base import Base
 from datetime import datetime, timezone
+from decimal import Decimal
 from typing import List, TYPE_CHECKING
 
 if TYPE_CHECKING:
@@ -16,9 +17,9 @@ class Proposal(Base):
     project_id: Mapped[int] = mapped_column(ForeignKey("projects.id"))
     freelancer_id: Mapped[int] = mapped_column(ForeignKey("users.id"))
     cover_letter: Mapped[str] = mapped_column(Text)
-    bid_amount: Mapped[float] = mapped_column(Float, nullable=False)  # Total bid amount for the project
+    bid_amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)  # Total bid amount for the project
     estimated_hours: Mapped[int] = mapped_column(Integer)
-    hourly_rate: Mapped[float] = mapped_column(Float)
+    hourly_rate: Mapped[Decimal] = mapped_column(Numeric(12, 2))
     availability: Mapped[str] = mapped_column(String(20))  # immediate, 1-2_weeks, 1_month, flexible
     attachments: Mapped[str] = mapped_column(Text, nullable=True)  # JSON string of attachment URLs
     status: Mapped[str] = mapped_column(String(20), default="submitted")  # submitted, accepted, rejected, withdrawn, draft
